US991053: Update secret format
https://internal.almoctane.com/ui/entity-navigation?p=131002/6001&entityType=work_item&id=991053

In Spring 6.2.0 the secret functionality has stopped working.

Previously, this method in CafConfigPropertySource was called with the string secret:GRAMMAR_DATABASE_PASSWORD:postgres:

public Object getProperty(final String name)
Now, with the latest version in Spring, it is no longer called with secret:GRAMMAR_DATABASE_PASSWORD:postgres, but called instead with just secret. As such the utility can no longer parse the secret key from it and get the secret.

Using a dot instead of a colon works. This results in the above method being called with secret.GRAMMAR_DATABASE_PASSWORD. There is no default value to parse now, but if the secret is not found, the default value is used as expected.